fix: create one to one from the profile view
