+++ wire-ios/Wire-iOS/Sources/Helpers/syncengine/UserType+Helpers.swift
@@ -17,7 +17,9 @@
 //
 
 import Foundation
-import WireDataModel
+import WireSyncEngine
+
+typealias ConversationCreatedBlock = (ZMConversation?) -> Void
 
 extension UserType {
 
@@ -31,6 +33,29 @@ extension UserType {
 
     var hasUntrustedClients: Bool {
         return allClients.contains { !$0.verified }
+    }
+
+    func createTeamOneToOneConversation(
+        in context: NSManagedObjectContext,
+        completion: @escaping ConversationCreatedBlock
+    ) {
+        guard
+            self.isTeamMember,
+            let user = self.materialize(in: context)
+        else {
+            return
+        }
+        let conversationService = ConversationService(context: context)
+
+        conversationService.createTeamOneToOneConversation(user: user) { result in
+            switch result {
+            case .success(let conversation):
+                completion(conversation)
+
+            case .failure(let error):
+                WireLogger.conversation.error("failed to create guest room: \(String(describing: error))")
+            }
+        }
     }
 
 }
